LPS-201076 SF, move BundleManagerUtil to existed util package. Also refer to WorkflowDefinitionManagerUtil
