+++ modules/apps/marketplace/marketplace-api/src/main/java/com/liferay/marketplace/util/BundleManagerUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.marketplace.bundle.util;
+package com.liferay.marketplace.util;
 
 import com.liferay.marketplace.bundle.BundleManager;
 import com.liferay.portal.kernel.module.service.Snapshot;
