+++ modules/apps/marketplace/marketplace-api/src/main/resources/com/liferay/marketplace/util/packageinfo
