+++ modules/apps/marketplace/marketplace-app-manager-web/src/main/java/com/liferay/marketplace/app/manager/web/internal/display/context/BaseAppManagerManagementToolbarDisplayContext.java
@@ -9,8 +9,8 @@ import com.liferay.frontend.taglib.clay.servlet.taglib.display.context.BaseManag
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
 import com.liferay.marketplace.app.manager.web.internal.constants.BundleStateConstants;
 import com.liferay.marketplace.app.manager.web.internal.util.MarketplaceAppManagerUtil;
-import com.liferay.marketplace.bundle.util.BundleManagerUtil;
 import com.liferay.marketplace.service.AppLocalServiceUtil;
+import com.liferay.marketplace.util.BundleManagerUtil;
 import com.liferay.petra.string.CharPool;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
